backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community     
chromadb
transformers
python-dotenv
pydantic
websockets
pandas
numpy
sentence-transformers
torch
scikit-learn
aiofiles
elasticsearch[async]==8.12.1
aiohttp